+++ typings/config.d.ts
@@ -1,10 +1,18 @@
 export const xmodules: {
+  core: boolean;
+  imgproc: boolean;
+  calib3d: boolean;
+  features2d: boolean;
+  io: boolean;
   dnn: boolean;
   face: boolean;
   text: boolean;
   tracking: boolean;
   xfeatures2d: boolean;
   ximgproc: boolean;
+  photo: boolean;
+  objdetect: boolean;
+  machinelearning: boolean;
 }
 
 export const version: {
@@ -12,3 +20,5 @@ export const version: {
   minor: number;
   revision: number;
 }
+
+export const modules: typeof cv.xmodules;